Move entropy test
