+++ tests/math/test_entropies_math.py
@@ -26,46 +26,6 @@ torch = pytest.importorskip("torch")
 jax = pytest.importorskip("jax")
 
 
-def expected_entropy_ising_xx(param):
-    """
-    Return the analytical entropy for the IsingXX.
-    """
-    eig_1 = (1 + np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)) / 2
-    eig_2 = (1 - np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)) / 2
-    eigs = [eig_1, eig_2]
-    eigs = [eig for eig in eigs if eig > 0]
-
-    expected_entropy = eigs * np.log(eigs)
-
-    expected_entropy = -np.sum(expected_entropy)
-    return expected_entropy
-
-
-def expected_entropy_grad_ising_xx(param):
-    """
-    Return the analytical gradient entropy for the IsingXX.
-    """
-    eig_1 = (1 + np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)) / 2
-    eig_2 = (1 - np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)) / 2
-    eigs = [eig_1, eig_2]
-    eigs = np.maximum(eigs, 1e-08)
-
-    grad_expected_entropy = -(
-        (np.log(eigs[0]) + 1)
-        * (np.sin(param / 2) ** 3 * np.cos(param / 2) - np.sin(param / 2) * np.cos(param / 2) ** 3)
-        / np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)
-    ) - (
-        (np.log(eigs[1]) + 1)
-        * (
-            np.sin(param / 2)
-            * np.cos(param / 2)
-            * (np.cos(param / 2) ** 2 - np.sin(param / 2) ** 2)
-        )
-        / np.sqrt(1 - 4 * np.cos(param / 2) ** 2 * np.sin(param / 2) ** 2)
-    )
-    return grad_expected_entropy
-
-
 class TestVonNeumannEntropy:
     """Tests for creating a density matrix from state vectors."""
 
@@ -83,8 +43,14 @@ class TestVonNeumannEntropy:
     @pytest.mark.parametrize("wires", single_wires_list)
     @pytest.mark.parametrize("state_vector,pure", state_vector)
     @pytest.mark.parametrize("check_state", check_state)
-    def test_state_vector_entropy_without_base(self, state_vector, wires, check_state, pure):
+    @pytest.mark.parametrize("interface", [None, "autograd", "jax", "tensorflow", "torch"])
+    def test_state_vector_entropy_without_base(
+        self, state_vector, wires, check_state, pure, interface
+    ):
         """Test entropy for different state vectors without base for log."""
+        if interface:
+            state_vector = qml.math.asarray(state_vector, like=interface)
+
         entropy = qml.math.vn_entropy(state_vector, wires, check_state=check_state)
 
         if pure:
@@ -97,8 +63,11 @@ class TestVonNeumannEntropy:
     @pytest.mark.parametrize("state_vector,pure", state_vector)
     @pytest.mark.parametrize("base", base)
     @pytest.mark.parametrize("check_state", check_state)
-    def test_state_vector_entropy(self, state_vector, wires, base, check_state, pure):
+    @pytest.mark.parametrize("interface", [None, "autograd", "jax", "tensorflow", "torch"])
+    def test_state_vector_entropy(self, state_vector, wires, base, check_state, pure, interface):
         """Test entropy for different state vectors."""
+        if interface:
+            state_vector = qml.math.asarray(state_vector, like=interface)
         entropy = qml.math.vn_entropy(state_vector, wires, base, check_state)
 
         if pure:
@@ -113,12 +82,17 @@ class TestVonNeumannEntropy:
         ([[1, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0]], True),
     ]
 
-    @pytest.mark.parametrize("wires", single_wires_list)
     @pytest.mark.parametrize("density_matrix,pure", density_matrices)
+    @pytest.mark.parametrize("wires", single_wires_list)
     @pytest.mark.parametrize("base", base)
     @pytest.mark.parametrize("check_state", check_state)
-    def test_density_matrices_entropy(self, density_matrix, wires, base, check_state, pure):
+    @pytest.mark.parametrize("interface", [None, "autograd", "jax", "tensorflow", "torch"])
+    def test_density_matrices_entropy(
+        self, density_matrix, pure, wires, base, check_state, interface
+    ):
         """Test entropy for different density matrices."""
+        if interface:
+            density_matrix = qml.math.asarray(density_matrix, like=interface)
         entropy = qml.math.vn_entropy(density_matrix, wires, base, check_state)
 
         if pure:
@@ -127,273 +101,6 @@ class TestVonNeumannEntropy:
             expected_entropy = np.log(2) / np.log(base)
 
         assert qml.math.allclose(entropy, expected_entropy)
-
-    parameters = np.linspace(0, 2 * np.pi, 20)
-
-    devices = ["default.qubit", "default.mixed"]
-
-    single_wires_list = [
-        [0],
-        [1],
-    ]
-
-    base = [2, np.exp(1), 10]
-
-    check_state = [True, False]
-
-    devices = ["default.qubit", "default.mixed"]
-    diff_methods = ["backprop", "finite-diff"]
-
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("device", devices)
-    @pytest.mark.parametrize("base", base)
-    def test_IsingXX_qnode_entropy(self, param, wires, device, base):
-        """Test entropy for a QNode numpy."""
-
-        dev = qml.device(device, wires=2)
-
-        @qml.qnode(dev)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        entropy = circuit_entropy(param)
-
-        expected_entropy = expected_entropy_ising_xx(param) / np.log(base)
-        assert qml.math.allclose(entropy, expected_entropy)
-
-    @pytest.mark.autograd
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("diff_method", diff_methods)
-    def test_IsingXX_qnode_entropy_grad(self, param, wires, base, diff_method):
-        """Test entropy for a QNode gradient with autograd."""
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, diff_method=diff_method)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        grad_entropy = qml.grad(circuit_entropy)(param)
-
-        # higher tolerance for finite-diff method
-        tol = 1e-8 if diff_method == "backprop" else 1e-5
-
-        grad_expected_entropy = expected_entropy_grad_ising_xx(param) / np.log(base)
-        assert qml.math.allclose(grad_entropy, grad_expected_entropy, atol=tol)
-
-    @pytest.mark.torch
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("device", devices)
-    @pytest.mark.parametrize("base", base)
-    def test_IsingXX_qnode_torch_entropy(self, param, wires, device, base):
-        """Test entropy for a QNode with torch interface."""
-        import torch
-
-        dev = qml.device(device, wires=2)
-
-        @qml.qnode(dev, interface="torch")
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        entropy = circuit_entropy(torch.tensor(param))
-
-        expected_entropy = expected_entropy_ising_xx(param) / np.log(base)
-
-        assert qml.math.allclose(entropy, expected_entropy)
-
-    @pytest.mark.torch
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("diff_method", diff_methods)
-    def test_IsingXX_qnode_entropy_grad_torch(self, param, wires, base, diff_method):
-        """Test entropy for a QNode gradient with torch."""
-        import torch
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, interface="torch", diff_method=diff_method)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        grad_expected_entropy = expected_entropy_grad_ising_xx(param) / np.log(base)
-
-        param = torch.tensor(param, dtype=torch.float64, requires_grad=True)
-        entropy = circuit_entropy(param)
-        entropy.backward()
-        grad_entropy = param.grad
-
-        # higher tolerance for finite-diff method
-        tol = 1e-8 if diff_method == "backprop" else 1e-5
-
-        assert qml.math.allclose(grad_entropy, grad_expected_entropy, atol=tol)
-
-    @pytest.mark.tf
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("device", devices)
-    @pytest.mark.parametrize("base", base)
-    def test_IsingXX_qnode_tf_entropy(self, param, wires, device, base):
-        """Test entropy for a QNode with tf interface."""
-        import tensorflow as tf
-
-        dev = qml.device(device, wires=2)
-
-        @qml.qnode(dev, interface="tf")
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        entropy = circuit_entropy(tf.Variable(param))
-
-        expected_entropy = expected_entropy_ising_xx(param) / np.log(base)
-
-        assert qml.math.allclose(entropy, expected_entropy)
-
-    @pytest.mark.tf
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("diff_method", diff_methods)
-    def test_IsingXX_qnode_entropy_grad_tf(self, param, wires, base, diff_method):
-        """Test entropy for a QNode gradient with tf."""
-        import tensorflow as tf
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, interface="tf", diff_method=diff_method)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        param = tf.Variable(param)
-        with tf.GradientTape() as tape:
-            entropy = circuit_entropy(param)
-
-        grad_entropy = tape.gradient(entropy, param)
-
-        grad_expected_entropy = expected_entropy_grad_ising_xx(param) / np.log(base)
-
-        # higher tolerance for finite-diff method
-        tol = 1e-8 if diff_method == "backprop" else 1e-5
-
-        assert qml.math.allclose(grad_entropy, grad_expected_entropy, atol=tol)
-
-    @pytest.mark.jax
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("device", devices)
-    @pytest.mark.parametrize("base", base)
-    def test_IsingXX_qnode_jax_entropy(self, param, wires, device, base):
-        """Test entropy for a QNode with jax interface."""
-        import jax.numpy as jnp
-
-        dev = qml.device(device, wires=2)
-
-        @qml.qnode(dev, interface="jax")
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        entropy = circuit_entropy(jnp.array(param))
-
-        expected_entropy = expected_entropy_ising_xx(param) / np.log(base)
-
-        assert qml.math.allclose(entropy, expected_entropy)
-
-    @pytest.mark.jax
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("diff_method", diff_methods)
-    def test_IsingXX_qnode_entropy_grad_jax(self, param, wires, base, diff_method):
-        """Test entropy for a QNode gradient with Jax."""
-        import jax
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, interface="jax", diff_method=diff_method)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        grad_entropy = jax.grad(circuit_entropy)(jax.numpy.array(param))
-        grad_expected_entropy = expected_entropy_grad_ising_xx(param) / np.log(base)
-
-        # higher tolerance for finite-diff method
-        tol = 1e-8 if diff_method == "backprop" else 1e-5
-
-        assert qml.math.allclose(grad_entropy, grad_expected_entropy, rtol=1e-04, atol=1e-05)
-
-    @pytest.mark.jax
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("device", devices)
-    def test_IsingXX_qnode_jax_jit_entropy(self, param, wires, base, device):
-        """Test entropy for a QNode with jax-jit interface."""
-        import jax
-        import jax.numpy as jnp
-
-        dev = qml.device(device, wires=2)
-
-        @qml.qnode(dev, interface="jax-jit")
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        entropy = jax.jit(circuit_entropy)(jnp.array(param))
-
-        expected_entropy = expected_entropy_ising_xx(param) / np.log(base)
-
-        assert qml.math.allclose(entropy, expected_entropy)
-
-    @pytest.mark.jax
-    @pytest.mark.parametrize("wires", single_wires_list)
-    @pytest.mark.parametrize("param", parameters)
-    @pytest.mark.parametrize("base", base)
-    @pytest.mark.parametrize("diff_method", diff_methods)
-    def test_IsingXX_qnode_entropy_grad_jax_jit(self, param, wires, base, diff_method):
-        """Test entropy for a QNode gradient with Jax-jit."""
-        import jax
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, interface="jax-jit", diff_method=diff_method)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=[0, 1])
-            return qml.vn_entropy(wires=wires, log_base=base)
-
-        grad_entropy = jax.jit(jax.grad(circuit_entropy))(jax.numpy.array(param))
-
-        grad_expected_entropy = expected_entropy_grad_ising_xx(param) / np.log(base)
-
-        assert qml.math.allclose(grad_entropy, grad_expected_entropy, rtol=1e-04, atol=1e-05)
-
-    @pytest.mark.parametrize("device", devices)
-    def test_qnode_entropy_no_custom_wires(self, device):
-        """Test that entropy cannot be returned with custom wires."""
-
-        dev = qml.device(device, wires=["a", 1])
-
-        @qml.qnode(dev)
-        def circuit_entropy(x):
-            qml.IsingXX(x, wires=["a", 1])
-            return qml.vn_entropy(wires=["a"])
-
-        with pytest.raises(
-            qml.QuantumFunctionError,
-            match="Returning the Von Neumann entropy is not supported when using custom wire labels",
-        ):
-            circuit_entropy(0.1)
 
 
 class TestMutualInformation:
